+++ src/components/accordion/accordion.component.js
@@ -3,7 +3,6 @@ import PropTypes from "prop-types";
 import propTypes from "@styled-system/prop-types";
 
 import useResizeObserver from "../../hooks/__internal__/useResizeObserver";
-import OptionsHelper from "../../utils/helpers/options-helper";
 import createGuid from "../../utils/helpers/guid";
 import Events from "../../utils/helpers/events";
 import {
@@ -17,7 +16,7 @@ import {
   StyledAccordionContent,
 } from "./accordion.style";
 import Button from "../button";
-import ValidationIcon from "../validations";
+import ValidationIcon from "../../__internal__/validations";
 
 const Accordion = React.forwardRef(
   (
@@ -213,7 +212,7 @@ Accordion.propTypes = {
   /** Sets icon type - accepted values: 'chevron_down' (default), 'dropdown' */
   iconType: PropTypes.oneOf(["chevron_down", "dropdown"]),
   /** Sets icon alignment - accepted values: 'left', 'right' (default) */
-  iconAlign: PropTypes.oneOf(OptionsHelper.alignBinary),
+  iconAlign: PropTypes.oneOf(["left", "right"]),
   /** Callback fired when expansion state changes, onChange(event: object, isExpanded: boolean) */
   onChange: PropTypes.func,
   /** Sets accordion title. Will render inside a h3 if set to a string */