+++ src/components/button-toggle/button-toggle-types.d.ts
@@ -1,0 +1,1 @@
+export type ButtonToggleIconSizes = "small" | "large";