+++ src/components/loader/loader.d.ts
@@ -1,9 +1,8 @@
 import { MarginProps } from "styled-system";
-import * as OptionsHelper from "../../utils/helpers/options-helper/options-helper";
 
 export interface LoaderProps extends MarginProps {
   /** Size of the loader. */
-  size?: OptionsHelper.SizesType;
+  size?: "small" | "large";
   /** Applies white color. */
   isInsideButton?: boolean;
   /** Applies slate color. Available only when isInsideButton is true. */