+++ src/components/message/message.stories.js
@@ -1,7 +1,6 @@
 import React, { useState } from "react";
 import { text, select, boolean } from "@storybook/addon-knobs";
 import { action } from "@storybook/addon-actions";
-import OptionsHelper from "../../utils/helpers/options-helper";
 import Button from "../button";
 import Message from "./message.component";
 
@@ -21,7 +20,11 @@ export default {
 
 export const Default = () => {
   const [isOpen, setIsOpen] = useState(true);
-  const variant = select("type", OptionsHelper.messages, "info");
+  const variant = select(
+    "type",
+    ["info", "error", "success", "warning"],
+    "info"
+  );
   const title = text("title");
   const id = text("id", "custom-id");
   const transparent = boolean("transparent", false);