+++ src/components/profile/profile.stories.js
@@ -2,7 +2,7 @@ import React from "react";
 import { text, select } from "@storybook/addon-knobs";
 
 import Profile from "./profile.component";
-import OptionsHelper from "../../utils/helpers/options-helper";
+import { PROFILE_SIZES } from "./profile.config";
 
 export default {
   title: "Profile/Test",
@@ -21,11 +21,7 @@ export default {
 export const Default = () => {
   const email = text("email", "johnsmith@sage.com");
   const initials = text("initials", "JS");
-  const size = select(
-    "size",
-    OptionsHelper.sizesPortrait,
-    OptionsHelper.sizesPortrait[0]
-  );
+  const size = select("size", PROFILE_SIZES, PROFILE_SIZES[0]);
   const name = text("name", "John Smith");
   const src = text("src", "");